backend/requirements.txt
gunicorn
django
django-debug-toolbar
djangorestframework
psycopg2-binary
python-dateutil
requests
vcrpy
